# server/docs/swagger.yaml
definitions:
  api.VerifyCaptchaRequest:
    properties:
      captchaCode:
        type: string
      captchaId:
        type: string
    required:
    - captchaCode
    - captchaId
    type: object
...
  github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.9.0
paths:
...
  /playmate/test-tool/verify-captcha:
    post:
      consumes:
      - application/json
      parameters:
      - description: 验证码信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.VerifyCaptchaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 验证成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 验证验证码（支持万能验证码123456）
      tags:
      - TestTool
  /playmate/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 类型 (avatar, image, voice, document)
        in: formData
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 上传文件
      tags:
      - Upload
  /playmate/upload/image:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 上传图片
      tags:
      - Upload
  /playmate/upload/voice:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 语音文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 上传语音
      tags:
      - Upload
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...